+++ core/src/components/button/bn-dislike-button.tsx
@@ -1,4 +1,4 @@
-import { Component, ComponentInterface, h, Prop, State, Watch } from "@stencil/core";
+import { Component, ComponentInterface, Event, EventEmitter, h, Prop, State, Watch } from "@stencil/core";
 import { assets } from "../../utils/assets";
 import { bindThis } from "../../utils/utils";
 import { $ImageToggleButton, imageToggleButtonRender } from "../util/bn-image-toggle-button.const";
@@ -19,6 +19,7 @@ export class BnButtonDislike implements ComponentInterface {
     this.updateCount();
   }
   @Prop({ mutable: true, reflect: true }) checked = false;
+  @Event() checkedChange!: EventEmitter<boolean>;
   @Prop() direction: $ImageToggleButton.Direction = "lr";
 
   @State() count_uncheck = 0;
@@ -27,6 +28,7 @@ export class BnButtonDislike implements ComponentInterface {
   @bindThis
   onChange(event: CustomEvent<boolean>) {
     this.checked = event.detail;
+    this.checkedChange.emit(event.detail);
   }
 
   connectedCallback() {