+++ angular/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bnqkl/web-component-angular
 
+## 1.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @bnqkl/web-component@1.3.0
+
 ## 1.0.6
 
 ### Patch Changes